chartrepo: support downloading JSON index files
This adds support for downloading `index.json` files from a Helm
repository endpoint.

The implementation works in such a way that it attempts to get the
index from a collection of URLs (`<URL>/index.json`,
`<URL>/index.yaml`; in that specific order) until it retrieves a
success response, after which it will attempt to read all data and
continue to load the index.

When it fails to get any index, it will return the collection of
getter errors as a single combined error.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>